requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
langchain==0.0.335
langchain-openai==0.0.2
chromadb==0.4.15
python-multipart==0.0.6
python-dotenv==1.0.0